[not verified] Moved SocialServiceIcon component from Jetpack Icons.js file to js-package/components
