+++ projects/js-packages/components/components/social-service-icon/Readme.md
@@ -1,0 +1,15 @@
+SocialServiceIcon
+========
+
+Component that renders different social service icons.
+
+#### How to use:
+
+```js
+<SocialServiceIcon serviceName={ serviceName } className="jetpack-publicize-gutenberg-social-icon" />
+```
+
+#### Props
+
+* `className`: String  the class name set on the SVG element.
+* `serviceName`: string - social service whose icon we want to render. Current available optiotns: facebook, twitter, linkedin, tumblr, google
