+++ projects/js-packages/components/index.jsx
@@ -42,3 +42,4 @@ export { default as ProductOffer, IconsCard } from './components/product-offer';
 export { default as Dialog } from './components/dialog';
 export { default as RecordMeterBar } from './components/record-meter-bar';
 export { getUserLocale, cleanLocale } from './lib/locale';
+export { SocialServiceIcon } from './components/social-service-icon';
