+++ projects/js-packages/components/package.json
@@ -6,6 +6,7 @@
 	"license": "GPL-2.0-or-later",
 	"dependencies": {
 		"@automattic/format-currency": "1.0.1",
+		"@wordpress/base-styles": "4.3.0",
 		"@wordpress/browserslist-config": "4.1.2",
 		"@wordpress/components": "19.10.0",
 		"@wordpress/compose": "5.5.0",
